+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### New features
 * Web: Cliking on the main graph in Dashboard redirects to the corresponding alerts
+* Web: Alerts preview when writing a condition
 
 ### Changes
 * Web: Updated all web packages + NodeJS (10->12)